src/main/java/ru/job4j/collection/StringCompare.java
package ru.job4j.collection;

import java.util.Comparator;

public class StringCompare implements Comparator<String> {
    @Override
    public int compare(String left, String right) {
        int size = Math.min(left.length(), right.length());
        for (int i = 0; i < size; i++) {
            char l = left.charAt(i);
            char r = right.charAt(i);
            int result = Character.compare(l, r);
            if (result != 0) {
                return result;
            }
        }
        return Integer.compare(left.length(), right.length());
    }
}
